Refactor authentication handling: move user info retrieval to TaakUserController and add DTOs for login requests and responses
