+++ backend/src/main/java/com/team8/taaks/dto/LoginRequest.java
@@ -1,0 +1,4 @@
+package com.team8.taaks.dto;
+
+public record LoginRequest(String username, String password) {
+}
