+++ backend/src/main/java/com/team8/taaks/dto/LoginResponse.java
@@ -1,0 +1,4 @@
+package com.team8.taaks.dto;
+
+public record LoginResponse(String token, UsersResponse user) {
+}
