+++ backend/src/main/java/com/team8/taaks/dto/UsersResponse.java
@@ -1,0 +1,4 @@
+package com.team8.taaks.dto;
+
+public record UsersResponse(String username, Long id) {
+}
